pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lam"
dynamic = ["version"]
description = "LAM (Locate Anything Model) — standalone geospatial tiling + grounded segmentation"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
dependencies = [
    "timm>=1.0.17",
    "numpy>=1.26",
    "scipy>=1.10.0",
    "tqdm",
    "ftfy==6.1.1",
    "regex",
    "iopath>=0.1.10",
    "typing_extensions",
    "huggingface_hub",
    "einops",
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "Pillow",
    "psutil>=5.9.0",
    "pycocotools>=2.0.6",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=4.0.0",
    "black>=25.9.0",
    "isort>=8.0.0",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.11.4",
]
serve = [
    "flask",
    "waitress",
    "opencv-python-headless",
    "json-logging",
    "orjson>=3.9",
]
geospatial = [
    "rasterio>=1.3.0",
    "shapely>=2.0.0",
    "pyproj>=3.5.0",
    "affine",
    "opencv-python-headless>=4.8.0",
    "matplotlib>=3.7.0",
]
# Matches upstream sam3.visualization_utils (plot_mask, COLORS, video overlays, COCO helpers).
viz = [
    "matplotlib>=3.7.0",
    "opencv-python-headless>=4.8.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "scikit-image>=0.22.0",
]

[project.scripts]
lam = "lam.cli:main"
lam-viz = "lam.geospatial.viz_geojson:main"

[project.urls]
Homepage = "https://github.com/drduhe/lam"

[tool.setuptools.packages.find]
where = ["."]
include = ["lam*"]
exclude = ["lam.perflib.tests*"]

[tool.setuptools.package-data]
lam = ["assets/bpe_simple_vocab_16e6.txt.gz", "agent/system_prompts/*.txt"]

[tool.setuptools.dynamic]
version = { attr = "lam.__version__" }

# Docstrings: Google style (summary line, then Args / Returns / Raises).
# Applied across lam/hosting, lam/geospatial, lam/prompt_mask, lam/perflib, and module
# banners for lam/train, lam/eval, lam/model, lam/agent (see scripts/add_missing_module_docstrings.py).
[tool.black]
line-length = 120
target-version = ["py313"]

[tool.isort]
profile = "black"
line_length = 120
known_first_party = ["lam"]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
# Forked upstream training/eval code triggers these; fix incrementally or tighten later.
ignore = [
    "E402",  # module import not at top (lazy / circular imports)
    "E721",  # type comparison with == (numpy / legacy)
    "E722",  # bare except
    "E731",  # lambda assignment
    "E741",  # ambiguous variable name (l, O, I)
    "F631",  # assert (a, b) tuple
    "F841",  # unused locals (WIP / scaffolding)
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = [
    "--strict-markers",
    "--cov=lam",
    "--cov-branch",
    "--cov-report=term-missing:skip-covered",
]
filterwarnings = [
    "ignore:Numba not found:UserWarning",
]

[tool.coverage.run]
branch = true
source = ["lam"]
omit = [
    "lam/perflib/tests/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "\\.\\.\\.",
    "raise NotImplementedError",
]
